vitaltrack-backend/requirements.txt
# VitalTrack Backend - Python Dependencies
# Production-ready FastAPI with async PostgreSQL

# Web Framework
fastapi[standard]==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
greenlet==3.1.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
argon2-cffi==23.1.0
bcrypt==4.2.1

# Validation & Serialization
pydantic==2.10.3
pydantic-settings==2.7.0
email-validator==2.2.0

# Date/Time
python-dateutil==2.9.0.post0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1

# Development
black==24.10.0
ruff==0.8.4
mypy==1.14.0

# Email
fastapi-mail==1.4.1
aiosmtplib>=2.0,<3.0

# Rate Limiting
slowapi==0.1.9

# Production
gunicorn==23.0.0
